react-app: rename recent proposal filename refs #148
